Split screening test into multiple files. Completed missing tests
