+++ openpharmacophore/screening/retrospective.py
@@ -1,6 +1,5 @@
 # OpenPharmacophore
 from openpharmacophore import Pharmacophore, LigandBasedPharmacophore, StructuredBasedPharmacophore
-from openpharmacophore.databases import chembl, pubchem
 from openpharmacophore.algorithms.alignment import apply_radii_to_bounds, transform_embeddings
 from openpharmacophore._private_tools.exceptions import BadShapeError, MissingParameters, OpenPharmacophoreValueError
 from openpharmacophore._private_tools.screening_arguments import check_virtual_screening_kwargs, is_3d_pharmacophore
@@ -17,12 +16,15 @@ from tqdm.auto import tqdm
 from collections import namedtuple
 from operator import itemgetter
 import os
-from typing import Callable, Tuple, List, Optional, TypeVar
+from typing import Tuple, List, Optional, TypeVar
 
-PharmacophoreType = TypeVar("PharmacophoreType", LigandBasedPharmacophore, 
-                    StructuredBasedPharmacophore, Pharmacophore, DataStructs.SparseBitVect)
+PharmacophoreType = TypeVar("PharmacophoreType", LigandBasedPharmacophore,
+                            StructuredBasedPharmacophore, Pharmacophore, DataStructs.SparseBitVect)
 
-class RetrospectiveScreening():
+MolScore = namedtuple("MolScore", ["score", "id", "mol"])
+
+
+class RetrospectiveScreening:
     """ Class for performing retrospective virtual screening. 
     
         This class expects molecules classified as actives and inactives. 
@@ -30,7 +32,7 @@ class RetrospectiveScreening():
 
     Parameters
     ----------
-    pharmacophore : openpharmacophore.Pharmacophore
+    pharmacophore : Pharmacophore
         The pharmacophore that will be used to screen the database. Can be a Pharmacophore, 
         StructuredBasedPharmacophore, LigandBasedPharmacophore or a fingerprint
 
@@ -47,7 +49,7 @@ class RetrospectiveScreening():
         Number of inactives.
 
     n_molecules: int
-        Number of molcules screened.
+        Number of molecules screened.
     
     n_fails : int
         Number of molecules that cannot be matched to the pharmacophore.
@@ -55,48 +57,34 @@ class RetrospectiveScreening():
     scoring_metric : str
         Metric used to score the molecules, how well they fit to the pharmacophore.
     
-    pharmacophore : openpharmacophore.Pharmacophore
+    pharmacophore : Pharmacophore
         The pharmacophore that will be used to screen the database. Can be a Pharmacophore, 
         StructuredBasedPharmacophore, LigandBasedPharmacophore or a fingerprint
 
+
+
     """
+
     def __init__(self, pharmacophore: PharmacophoreType, **kwargs) -> None:
-        
+
         if is_3d_pharmacophore(pharmacophore):
             self.scoring_metric = "SSD"
             self._screen_fn = self._align_molecules
-        elif isinstance(pharmacophore, DataStructs.SparseBitVect): # For pharmacophore fingerprints
+        elif isinstance(pharmacophore, DataStructs.SparseBitVect):  # For pharmacophore fingerprints
             self.scoring_metric = "Similarity"
             self.similarity_fn, _ = check_virtual_screening_kwargs(**kwargs)
             self._factory = Gobbi_Pharm2D.factory
             self._screen_fn = self._fingerprint_similarity
         else:
             raise TypeError("pharmacophore must be of type Pharmacophore, StructuredBasedPharmacophore, "
-                "LigandBasedPharmacophore, or rdkit.DataStructs.SparseBitVect")
-        
+                            "LigandBasedPharmacophore, or rdkit.DataStructs.SparseBitVect")
+
         self.n_actives = 0
         self.n_inactives = 0
         self.bioactivities = None
         self.molecules = []
         self.n_molecules = 0
         self.pharmacophore = pharmacophore
-             
-    def from_chembl_target_id(self, target_id: str, pIC50_threshold: float = 6.3) -> None:
-        """ Retrospective screening from bioactivity data fetched from chembl.
-           
-           Parameters
-           ----------
-           target_id : str
-                ChemBl target id.
-           
-           pIC50_threshold : float, default=6.3
-                The cuttoff value from which a molecule is considered active.
-           
-           """
-        smiles, activity = chembl.get_training_data(target_id, pIC50_threshold)
-        
-        self.db = "PubChem"
-        self.from_bioactivity_data(smiles, activity)
 
     def from_bioactivity_data(self, smiles: List[Tuple[int, str]], activity: np.ndarray) -> None:
         """ Retrospective screening from a set of molecules classified as active or inactive.
@@ -124,9 +112,9 @@ class RetrospectiveScreening():
         self.bioactivities = activity
 
         molecules = []
-        for id, smi in smiles:
+        for id_, smi in smiles:
             mol = Chem.MolFromSmiles(smi)
-            mol.SetProp("_Name", str(id))
+            mol.SetProp("_Name", str(id_))
             molecules.append(mol)
 
         if self.scoring_metric == "Similarity":
@@ -135,19 +123,6 @@ class RetrospectiveScreening():
             self._align_molecules(molecules)
         else:
             raise NotImplementedError
-
-    def from_pubchem_bioassay_id(self, bioassay_id: int) -> None:
-        """ Retrospective screening from a pubchem bioassay.
-
-            Parameters
-            ----------
-            bioassay_id : int
-                PubChem bioassay id. 
-        """
-        pubchem_client = pubchem.PubChem()
-        smiles, activity = pubchem_client.get_assay_training_data(bioassay_id)
-        self.db = "Pubchem"
-        self.from_bioactivity_data(smiles, activity)
 
     def confusion_matrix(self, threshold: Optional[float] = None) -> np.ndarray:
         """ Compute a confusion matrix
@@ -166,15 +141,18 @@ class RetrospectiveScreening():
         """
         if self.scoring_metric == "Similarity" and threshold is None:
             raise MissingParameters("Expected a threshold value.")
-        
+
         if self.scoring_metric == "SSD":
+            # TODO: is threshold value of 0 correct? An SSD of 0 means that
+            #  the pharmacophore aligns perfectly with the ligand. However this is really
+            #  unlikely. We should establish a higher threshold
             threshold = 0.0
-            
+
         true_positives = 0
         true_negatives = 0
         false_positives = 0
         false_negatives = 0
-        
+
         for ii, mol in enumerate(self.molecules):
             if self.bioactivities[ii] == 1 and mol.score > threshold:
                 true_positives += 1
@@ -184,15 +162,15 @@ class RetrospectiveScreening():
                 false_positives += 1
             elif self.bioactivities[ii] == 0 and mol.score <= threshold:
                 true_negatives += 1
-        
+
         cf_matrix = np.array([[true_positives, false_positives],
                               [false_negatives, true_negatives]])
-        
+
         assert np.sum(cf_matrix, axis=None) == self.n_molecules
-        
+
         return cf_matrix
 
-    def auc(self) -> Callable[[np.ndarray, np.ndarray], float]:
+    def auc(self) -> float:
         """ Calculate ROC area under the curve.
         
             Returns
@@ -209,8 +187,8 @@ class RetrospectiveScreening():
 
         return self._get_auc(scores, self.bioactivities)
 
-    def roc_plot(self, ax: Optional[plt.Axes] = None, label: str = "", 
-                random_line: bool = True) -> plt.Axes:
+    def roc_plot(self, ax: Optional[plt.Axes] = None, label: str = "",
+                 random_line: bool = True) -> plt.Axes:
         """ Plot the ROC curve. 
         
             Parameters
@@ -221,10 +199,13 @@ class RetrospectiveScreening():
             random_line : bool, default=True
                 Whether to plot the line corresponding to a random classifier.
 
+            label : str, default=""
+                The label of the ROC curve
+
             Returns
             ----------
             ax : matplotlib.axes._subplots.AxesSubplot, optional (Default = None)
-                An axes object whith the plot.
+                An axes object with the plot.
 
             References
             ----------
@@ -242,11 +223,11 @@ class RetrospectiveScreening():
         ax.plot(fpr, tpr, label=label)
         if random_line:
             ax.plot([0, 1], [0, 1], color="black", linestyle="dashed", label="Random")
-        
+
         ax.set_ylabel("Sensitivity")
         ax.set_xlabel("1 - Specificity")
         if label or random_line:
-            ax.legend()   
+            ax.legend()
 
         return ax
 
@@ -265,12 +246,11 @@ class RetrospectiveScreening():
         """
         if percentage < 0 or percentage > 100:
             raise OpenPharmacophoreValueError("percentage must be a number between 0 and 100")
-        
+
         scores = [x[0] for x in self.molecules]
         scores = np.array(scores)
 
         return self._calculate_enrichment_factor(scores, self.bioactivities, percentage)
-       
 
     def ideal_enrichment_factor(self, percentage: float) -> float:
         """ Calculate ideal enrichment factor for the x% of the screened database 
@@ -283,8 +263,8 @@ class RetrospectiveScreening():
             Returns
             -------
             float
-                The idal enrichment factor"""
-    
+                The ideal enrichment factor"""
+
         percentage = percentage / 100
 
         n_molecules = self.bioactivities.shape[0]
@@ -293,8 +273,8 @@ class RetrospectiveScreening():
             return (100 / ratio_actives) * percentage
         else:
             return 100.0
-    
-    def enrichment_plot(self, ax: Optional[plt.Axes] = None, label: str = "", 
+
+    def enrichment_plot(self, ax: Optional[plt.Axes] = None, label: str = "",
                         random_line: bool = True, ideal: bool = False) -> plt.Axes:
         """ Create an enrichment plot 
             
@@ -306,20 +286,22 @@ class RetrospectiveScreening():
             random_line : bool, default=True
                 Whether to plot the line corresponding to a random classifier.
 
-            ideal : bool, defaul=False
-                Whether to plot the ideal enrichmnent curve
+            ideal : bool, default=False
+                Whether to plot the ideal enrichment curve
+
+             label : str, default=""
+                The label of the enrichment curve
             
             Returns
             ----------
             ax : matplotlib.axes._subplots.AxesSubplot, optional (Default = None)
-                An axes object whith the plot.
+                An axes object with the plot.
         
         """
         scores = [x[0] for x in self.molecules]
         scores = np.array(scores)
-        
+
         screened_percentage, percentage_actives_found = self._enrichment_data(scores, self.bioactivities)
-        n_molecules = self.bioactivities.shape[0]
 
         if ax is None:
             fig, ax = plt.subplots()
@@ -330,13 +312,29 @@ class RetrospectiveScreening():
             n_molecules = self.bioactivities.shape[0]
             ratio_actives = self.n_actives / n_molecules
             ax.plot([0, ratio_actives, 1], [0, 1, 1], color="red", linestyle="dashed", label="Ideal")
-        
+
         ax.set_xlabel("% Database Screened")
-        ax.set_ylabel("% Actives")   
+        ax.set_ylabel("% Actives")
         ax.legend()
 
         return ax
-    
+
+    @staticmethod
+    def _sort_scores_and_labels_for_roc(scores: np.ndarray,
+                                        labels: np.ndarray) -> Tuple[np.ndarray, np.ndarray]:
+        """ Sorts scores and labels needed for ROC an AUC calculations.
+        """
+        if len(scores.shape) > 1 or len(labels.shape) > 1:
+            raise BadShapeError("scores and labels must be an array of rank one.")
+        if scores.shape[0] != labels.shape[0]:
+            raise BadShapeError("scores and labels must have the same shape.")
+        # Sort scores in descending order and then sort labels
+        indices = np.argsort(scores)[::-1]
+        scores = np.sort(scores)[::-1]
+        labels = labels[indices]
+
+        return scores, labels
+
     @staticmethod
     def _roc_points(scores: np.ndarray, labels: np.ndarray) -> Tuple[List[float], List[float]]:
         """ Calculate points to plot an ROC curve.
@@ -357,46 +355,104 @@ class RetrospectiveScreening():
             true_positive_rate : list of float
                 A list with the values of true positive rate.
         """
-        if len(scores.shape) > 1 or len(labels.shape) > 1:
-            raise BadShapeError("scores and labels must be an array of rank one.")
-        if scores.shape[0] != labels.shape[0]:
-            raise BadShapeError("scores and labels must have the same shape.")
-        # Sort scores in descending order and then sort labels
-        indices = np.argsort(scores)[::-1]
-        scores = np.sort(scores)[::-1] 
-        labels = labels[indices]
+        scores, labels = RetrospectiveScreening._sort_scores_and_labels_for_roc(scores, labels)
 
         n_positives = np.sum(labels)
         n_negatives = labels.shape[0] - n_positives
-        
+
+        score_prev = -10000000
+
         false_positives = 0
         true_positives = 0
         false_positive_rate = []
         true_positive_rate = []
-        score_prev = -10000000
 
         # Calculate points for the ROC plot
-        i = 0
-        while i < labels.shape[0]:
+        ii = 0
+        while ii < labels.shape[0]:
 
-            if scores[i] != score_prev:
-                false_positive_rate.append(false_positives/n_negatives)
-                true_positive_rate.append(true_positives/n_positives)
-                score_prev = scores[i]
+            if scores[ii] != score_prev:
+                false_positive_rate.append(false_positives / n_negatives)
+                true_positive_rate.append(true_positives / n_positives)
+                score_prev = scores[ii]
 
-            if labels[i] == 1:
+            if labels[ii] == 1:
                 true_positives += 1
             else:
                 false_positives += 1
 
-            i += 1
+            ii += 1
 
         # Append point (1, 1)
-        false_positive_rate.append(false_positives/n_negatives)
-        true_positive_rate.append(true_positives/n_positives)
-        
+        false_positive_rate.append(false_positives / n_negatives)
+        true_positive_rate.append(true_positives / n_positives)
+
         return false_positive_rate, true_positive_rate
-    
+
+    @staticmethod
+    def _get_auc(scores: np.ndarray, labels: np.ndarray) -> float:
+        """ Compute the area under the ROC curve.
+
+            Parameters
+            ----------
+            scores : np.ndarray of shape(n_samples,)
+                The score of each sample.
+
+            labels : np.ndarray of shape(n_samples,)
+                The ground truth.
+
+            Returns
+            --------
+            area : float
+                The area under the curve.
+        """
+        scores, labels = RetrospectiveScreening._sort_scores_and_labels_for_roc(scores, labels)
+
+        n_positives = int(np.sum(labels))
+        n_negatives = labels.shape[0] - n_positives
+
+        false_positives = 0
+        true_positives = 0
+        false_pos_prev = 0
+        true_pos_prev = 0
+
+        area = 0
+        score_prev = -10000000
+
+        ii = 0
+        while ii < labels.shape[0]:
+
+            if scores[ii] != score_prev:
+                area += RetrospectiveScreening._trapezoid_area(false_positives, false_pos_prev,
+                                                               true_positives, true_pos_prev)
+                score_prev = scores[ii]
+                false_pos_prev = false_positives
+                true_pos_prev = true_positives
+
+            if labels[ii] == 1:
+                true_positives += 1
+            else:
+                false_positives += 1
+
+            ii += 1
+
+        area += RetrospectiveScreening._trapezoid_area(n_negatives, false_pos_prev, n_positives, true_pos_prev)
+        # Scale area from n_negatives * n_positives onto the unit square
+        area = area / (n_negatives * n_positives)
+
+        assert 0 <= area <= 1
+
+        return area
+
+    @staticmethod
+    def _trapezoid_area(x1: float, x2: float, y1: float, y2: float) -> float:
+        """ Calculate the area of a trapezoid.
+        """
+        base = abs(x1 - x2)
+        # average height
+        height = abs(y1 + y2) / 2
+        return base * height
+
     @staticmethod
     def _enrichment_data(scores: np.ndarray, labels: np.ndarray) -> Tuple[List[float], List[float]]:
         """ Get enrichment data necessary for enrichment plot and enrichment factor calculation.
@@ -418,14 +474,7 @@ class RetrospectiveScreening():
                 The percentage of actives found.
 
         """
-        if len(scores.shape) > 1 or len(labels.shape) > 1:
-            raise BadShapeError("scores and labels must be an array of rank one.")
-        if scores.shape[0] != labels.shape[0]:
-            raise BadShapeError("scores and labels must have the same shape.")
-        # Sort scores in descending order and then sort labels
-        indices = np.argsort(scores)[::-1]
-        scores = np.sort(scores)[::-1] 
-        bioactivities = labels[indices]
+        scores, bioactivities = RetrospectiveScreening._sort_scores_and_labels_for_roc(scores, labels)
 
         n_molecules = bioactivities.shape[0]
         n_actives = np.sum(bioactivities)
@@ -439,9 +488,9 @@ class RetrospectiveScreening():
                 actives_counter += 1
             percentage_actives_found.append(actives_counter / n_actives)
             screened_percentage.append((ii + 1) / n_molecules)
-        
+
         return screened_percentage, percentage_actives_found
-    
+
     @staticmethod
     def _calculate_enrichment_factor(scores: np.ndarray, labels: np.ndarray, percentage: float) -> float:
         """ Calculate enrichment factor for the x% of the screened database 
@@ -469,78 +518,7 @@ class RetrospectiveScreening():
         indices_screen_per = screened_percentage <= percentage / 100
         max_enrichment_idx = np.argsort(screened_percentage[indices_screen_per])[-1]
 
-        return percentage_actives_found[max_enrichment_idx] * 100    
-    
-    @staticmethod
-    def _get_auc(scores: np.ndarray, labels: np.ndarray) -> float:
-        """ Compute the area under the ROC curve.
-        
-            Parameters
-            ----------
-            scores : np.ndarray of shape(n_samples,)
-                The score of each sample.
-            
-            labels : np.ndarray of shape(n_samples,)
-                The ground truth.
-
-            Returns
-            --------
-            area : float
-                The area under the curve.
-        """
-        if len(scores.shape) > 1 or len(labels.shape) > 1:
-            raise BadShapeError("scores and labels must be an array of rank one.")
-        if scores.shape[0] != labels.shape[0]:
-            raise BadShapeError("scores and labels must have the same shape.")
-        # Sort scores in descending order and then sort labels
-        indices = np.argsort(scores)[::-1]
-        scores = np.sort(scores)[::-1] 
-        labels = labels[indices]
-
-        n_positives = np.sum(labels)
-        n_negatives = labels.shape[0] - n_positives
-
-        false_positives = 0
-        true_positives = 0
-        false_pos_prev = 0
-        true_pos_prev = 0
-
-        area = 0
-        score_prev = -10000000
-
-        i = 0
-        while i < labels.shape[0]:
-
-            if scores[i] != score_prev:
-                area += RetrospectiveScreening._trapezoid_area(false_positives, false_pos_prev, 
-                                        true_positives, true_pos_prev)
-                score_prev = scores[i]
-                false_pos_prev = false_positives
-                true_pos_prev = true_positives
-            
-            if labels[i] == 1:
-                true_positives += 1
-            else:
-                false_positives += 1
-
-            i += 1
-
-        area += RetrospectiveScreening._trapezoid_area(n_negatives, false_pos_prev, n_positives, true_pos_prev)
-        # Scale area from n_negatives * n_positives onto the unit square
-        area = area / (n_negatives * n_positives)
-
-        assert area >= 0 and area <= 1
-        
-        return area
-    
-    @staticmethod
-    def _trapezoid_area(x1: float, x2: float, y1: float, y2: float) -> float:
-        """ Calculate the area of a trapezoid.
-        """
-        base = abs(x1 - x2)
-        # average height
-        height = abs(y1 + y2) / 2
-        return base * height
+        return percentage_actives_found[max_enrichment_idx] * 100
 
     def _align_molecules(self, molecules: List[Chem.Mol]) -> None:
         """ Align a list of molecules to a given pharmacophore.
@@ -556,20 +534,18 @@ class RetrospectiveScreening():
         rdkit_pharmacophore, radii = self.pharmacophore.to_rdkit()
         apply_radii_to_bounds(radii, rdkit_pharmacophore)
 
-        fdef = os.path.join(RDConfig.RDDataDir,'BaseFeatures.fdef')
+        fdef = os.path.join(RDConfig.RDDataDir, 'BaseFeatures.fdef')
         featFactory = ChemicalFeatures.BuildFeatureFactory(fdef)
-        
-        MolScore = namedtuple("MolScore", ["score", "id", "mol"])
-        
+
         for mol in tqdm(molecules):
 
             bounds_matrix = rdDistGeom.GetMoleculeBoundsMatrix(mol)
             can_match, all_matches = EmbedLib.MatchPharmacophoreToMol(mol, featFactory, rdkit_pharmacophore)
             if can_match:
-                failed, _ , matched_mols, _ = EmbedLib.MatchPharmacophore(all_matches, 
-                                                                          bounds_matrix,
-                                                                          rdkit_pharmacophore, 
-                                                                          useDownsampling=True)
+                failed, _, matched_mols, _ = EmbedLib.MatchPharmacophore(all_matches,
+                                                                         bounds_matrix,
+                                                                         rdkit_pharmacophore,
+                                                                         useDownsampling=True)
                 if failed:
                     matched_mol = MolScore(0.0, mol.GetProp("_Name"), mol)
                     self.molecules.append(matched_mol)
@@ -579,24 +555,25 @@ class RetrospectiveScreening():
                 self.molecules.append(matched_mol)
                 continue
             atom_match = [list(x.GetAtomIds()) for x in matched_mols]
-            
+
             try:
                 mol_H = Chem.AddHs(mol)
                 _, embeddings, _ = EmbedLib.EmbedPharmacophore(mol_H, atom_match, rdkit_pharmacophore, count=10)
             except:
                 continue
-            
-            SSDs = transform_embeddings(rdkit_pharmacophore, embeddings, atom_match) 
+
+            SSDs = transform_embeddings(rdkit_pharmacophore, embeddings, atom_match)
             if len(SSDs) == 0:
+                # TODO: Should we give a score of zero to molecules that don't match?
                 matched_mol = MolScore(0.0, mol.GetProp("_Name"), mol)
                 self.molecules.append(matched_mol)
                 continue
             best_fit_index = min(enumerate(SSDs), key=itemgetter(1))[0]
-            
+
             score = 1 / SSDs[best_fit_index]
             matched_mol = MolScore(score, mol.GetProp("_Name"), embeddings[best_fit_index])
             self.molecules.append(matched_mol)
-    
+
     def _fingerprint_similarity(self, molecules: List[Chem.Mol]) -> None:
         """ Compute fingerprints and similarity values for a list of molecules. 
 
@@ -606,16 +583,15 @@ class RetrospectiveScreening():
             List of molecules whose similarity to the pharmacophoric fingerprint will be calculated.
 
         """
-       
+
         self.n_molecules = len(molecules)
-        MolScore = namedtuple("MolScore", ["score", "id", "mol"])
-        
+
         for mol in tqdm(molecules):
             fingerprint = Gen2DFingerprint(mol, self._factory)
             similarity = self.similarity_fn(self.pharmacophore, fingerprint)
             mol_id = mol.GetProp("_Name")
             matched_mol = MolScore(similarity, mol_id, mol)
             self.molecules.append(matched_mol)
-          
+
     def __repr__(self) -> str:
         return f"{self.__class__.__name__}(n_molecules={self.n_molecules})"
